backend/requirements.txt
flask
numpy
pandas
scikit-learn
ollama
langchain
langchain-huggingface
transformers
huggingface-hub
python-dotenv
scikit-learn
langchain_community
langserve
fastapi
langchainhub
sentence_transformers
langchain-objectbox
flask-cors
opencv-python 
mediapipe